+++ src/components/overlay/Overlay.js
@@ -51,18 +51,11 @@ class Overlay extends PureComponent {
     }
   };
 
-  handleClick = (event) => {
-    event.stopPropagation();
-    this.props.onClick?.(event);
-  };
-
   handleMouseDown = (event) => {
-    event.stopPropagation();
     this.clickOriginRef.current = event.target;
   };
 
   handleMouseUp = (event) => {
-    event.stopPropagation();
     // Only register clicks outside of the children
     if (
       this.props.onOverlayClick &&
@@ -90,7 +83,6 @@ class Overlay extends PureComponent {
               onKeyDown={this.handleEscKey}
               onMouseDown={this.handleMouseDown}
               onMouseUp={this.handleMouseUp}
-              onClick={this.handleClick}
               className={cx(
                 theme['overlay'],
                 theme[backdrop],